// src/types/ngo.ts
export interface Ngo {
  name: string;
  description?: string;
  email: string;
  phone?: string;
  cnpj?: string;
  city?: string;
  website?: string;
  instagram?: string;
  facebook?: string;
  tiktok?: string;
  adoptionForm: string;
  sponsorshipForm?: string;
  temporaryHomeForm?: string;
  claimForm?: string;
}